+++ tests/test_classifiers.py
@@ -119,39 +119,46 @@ class TestNaiveBayesClassifier(unittest.TestCase):
         assert_equal(cl.train_features[0][1], 'positive')
 
     def test_init_with_csv_file(self):
-        cl = NaiveBayesClassifier(CSV_FILE, format="csv")
+        with open(CSV_FILE) as fp:
+            cl = NaiveBayesClassifier(fp, format="csv")
         assert_equal(cl.classify("I feel happy this morning"), 'pos')
         training_sentence = cl.train_set[0][0]
         assert_true(isinstance(training_sentence, unicode))
 
     def test_init_with_csv_file_without_format_specifier(self):
-        cl = NaiveBayesClassifier(CSV_FILE)
+        with open(CSV_FILE) as fp:
+            cl = NaiveBayesClassifier(fp)
         assert_equal(cl.classify("I feel happy this morning"), 'pos')
         training_sentence = cl.train_set[0][0]
         assert_true(isinstance(training_sentence, unicode))
 
     def test_init_with_json_file(self):
-        cl = NaiveBayesClassifier(JSON_FILE, format="json")
+        with open(JSON_FILE) as fp:
+            cl = NaiveBayesClassifier(fp, format="json")
         assert_equal(cl.classify("I feel happy this morning"), 'pos')
         training_sentence = cl.train_set[0][0]
         assert_true(isinstance(training_sentence, unicode))
 
     def test_init_with_json_file_without_format_specifier(self):
-        cl = NaiveBayesClassifier(JSON_FILE)
+        with open(JSON_FILE) as fp:
+            cl = NaiveBayesClassifier(fp)
         assert_equal(cl.classify("I feel happy this morning"), 'pos')
         training_sentence = cl.train_set[0][0]
         assert_true(isinstance(training_sentence, unicode))
 
     def test_accuracy_on_a_csv_file(self):
-        a = self.classifier.accuracy(CSV_FILE)
-        assert_true(isinstance(a, float))
+        with open(CSV_FILE) as fp:
+            a = self.classifier.accuracy(fp)
+        assert_equal(type(a), float)
 
     def test_accuracy_on_json_file(self):
-        a = self.classifier.accuracy(JSON_FILE)
-        assert_true(isinstance(a, float))
+        with open(CSV_FILE) as fp:
+            a = self.classifier.accuracy(fp)
+        assert_equal(type(a), float)
 
     def test_init_with_tsv_file(self):
-        cl = NaiveBayesClassifier(TSV_FILE)
+        with open(TSV_FILE) as fp:
+            cl = NaiveBayesClassifier(fp)
         assert_equal(cl.classify("I feel happy this morning"), 'pos')
         training_sentence = cl.train_set[0][0]
         assert_true(isinstance(training_sentence, unicode))