+++ tests/test_formats.py
@@ -7,7 +7,6 @@ from nose.tools import *  # PEP8 asserts
 from textblob import formats
 from textblob.compat import unicode
 
-logging.basicConfig(level=logging.DEBUG)
 HERE = os.path.abspath(os.path.dirname(__file__))
 CSV_FILE = os.path.join(HERE, 'data.csv')
 JSON_FILE = os.path.join(HERE, "data.json")
@@ -19,11 +18,13 @@ class TestFormats(unittest.TestCase):
         pass
 
     def test_detect_csv(self):
-        format = formats.detect(CSV_FILE)
+        with open(CSV_FILE) as fp:
+            format = formats.detect(fp)
         assert_equal(format, formats.CSV)
 
     def test_detect_json(self):
-        format = formats.detect(JSON_FILE)
+        with open(JSON_FILE) as fp:
+            format = formats.detect(fp)
         assert_equal(format, formats.JSON)
 
     def test_available(self):
@@ -47,7 +48,8 @@ class TestDelimitedFormat(unittest.TestCase):
 class TestCSV(unittest.TestCase):
 
     def test_read_from_filename(self):
-        data = formats.CSV(CSV_FILE)
+        with open(CSV_FILE) as fp:
+            data = formats.CSV(fp)
 
     def test_detect(self):
         with open(CSV_FILE, 'r') as fp:
@@ -59,8 +61,9 @@ class TestCSV(unittest.TestCase):
 
 class TestTSV(unittest.TestCase):
 
-    def test_read_from_filename(self):
-        data = formats.TSV(TSV_FILE)
+    def test_read_from_file_object(self):
+        with open(TSV_FILE) as fp:
+            data = formats.TSV(fp)
 
     def test_detect(self):
         with open(TSV_FILE, 'r') as fp:
@@ -73,8 +76,9 @@ class TestTSV(unittest.TestCase):
 
 class TestJSON(unittest.TestCase):
 
-    def test_read_from_filename(self):
-        formats.JSON(JSON_FILE)
+    def test_read_from_file_object(self):
+        with open(JSON_FILE) as fp:
+            formats.JSON(fp)
 
     def test_detect(self):
         with open(JSON_FILE, 'r') as fp:
@@ -85,8 +89,8 @@ class TestJSON(unittest.TestCase):
             assert_false(formats.JSON.detect(stream))
 
     def test_to_iterable(self):
-        d = formats.JSON(JSON_FILE)
-        logging.debug(d.dict)
+        with open(JSON_FILE) as fp:
+            d = formats.JSON(fp)
         data = d.to_iterable()
         first = data[0]
         text, label = first[0], first[1]