+++ textblob/classifiers.py
@@ -39,7 +39,7 @@ import nltk
 from textblob.tokenizers import word_tokenize
 from textblob.compat import basestring
 import textblob.formats as formats
-from textblob.utils import strip_punc
+from textblob.utils import strip_punc, is_filelike
 from textblob.decorators import cached_property
 
 ### Basic feature extractors ###
@@ -100,7 +100,7 @@ class BaseClassifier(object):
     a ``classifier`` property.
 
     :param train_set: The training set, either a list of tuples of the form
-        ``(text, classification)`` or a filename. ``text`` may be either
+        ``(text, classification)`` or a file pointer. ``text`` may be either
         a string or an iterable.
     :param callable feature_extractor: A feature extractor function that takes one or
         two arguments: ``document`` and ``train_set``.
@@ -113,7 +113,7 @@ class BaseClassifier(object):
 
     def __init__(self, train_set, feature_extractor=basic_extractor, format=None):
         self.feature_extractor = feature_extractor
-        if isinstance(train_set, basestring):  # train_set is a filename
+        if is_filelike(train_set):
             self.train_set = self._read_data(train_set, format)
         else:  # train_set is a list of tuples
             self.train_set = train_set
@@ -134,19 +134,19 @@ class BaseClassifier(object):
 
     @cached_property
     def classifier(self):
-        '''The classifier object.'''
+        """The classifier object."""
         raise NotImplementedError('Must implement the "classifier" property.')
 
     def classify(self, text):
-        '''Classifies a string of text.'''
+        """Classifies a string of text."""
         raise NotImplementedError('Must implement a "classify" method.')
 
     def train(self, labeled_featureset):
-        '''Trains the classifier.'''
+        """Trains the classifier."""
         raise NotImplementedError('Must implement a "train" method.')
 
     def labels(self):
-        '''Returns an iterable containing the possible labels.'''
+        """Returns an iterable containing the possible labels."""
         raise NotImplementedError('Must implement a "labels" method.')
 
     def extract_features(self, text):
@@ -162,7 +162,6 @@ class BaseClassifier(object):
 
 
 class NLTKClassifier(BaseClassifier):
-
     """An abstract class that wraps around the nltk.classify module.
 
     Expects that descendant classes include a class variable ``nltk_class``
@@ -172,7 +171,6 @@ class NLTKClassifier(BaseClassifier):
 
         class MyClassifier(NLTKClassifier):
             nltk_class = nltk.classify.svm.SvmClassifier
-
     """
 
     #: The NLTK class to be wrapped. Must be a class within nltk.classify
@@ -190,7 +188,7 @@ class NLTKClassifier(BaseClassifier):
 
     @cached_property
     def classifier(self):
-        '''The classifier.'''
+        """The classifier."""
         try:
             return self.train()
         except AttributeError:  # nltk_class has not been defined
@@ -198,7 +196,7 @@ class NLTKClassifier(BaseClassifier):
                             " variable that is not None.")
 
     def train(self, *args, **kwargs):
-        '''Train the classifier with a labeled feature set and return
+        """Train the classifier with a labeled feature set and return
         the classifier. Takes the same arguments as the wrapped NLTK class.
         This method is implicitly called when calling ``classify`` or
         ``accuracy`` methods and is included only to allow passing in arguments
@@ -207,7 +205,7 @@ class NLTKClassifier(BaseClassifier):
         .. versionadded:: 0.6.2
 
         :rtype: A classifier
-        '''
+        """
         try:
             self.classifier = self.nltk_class.train(self.train_features,
                                                     *args, **kwargs)
@@ -217,27 +215,27 @@ class NLTKClassifier(BaseClassifier):
                             " variable that is not None.")
 
     def labels(self):
-        '''Return an iterable of possible labels.'''
+        """Return an iterable of possible labels."""
         return self.classifier.labels()
 
     def classify(self, text):
-        '''Classifies the text.
+        """Classifies the text.
 
         :param str text: A string of text.
-        '''
+        """
         text_features = self.extract_features(text)
         return self.classifier.classify(text_features)
 
     def accuracy(self, test_set, format=None):
-        '''Compute the accuracy on a test set.
+        """Compute the accuracy on a test set.
 
         :param test_set: A list of tuples of the form ``(text, label)``, or a
-            filename.
+            file pointer.
         :param format: If ``test_set`` is a filename, the file format, e.g.
             ``"csv"`` or ``"json"``. If ``None``, will attempt to detect the
             file format.
-        '''
-        if isinstance(test_set, basestring):  # test_set is a filename
+        """
+        if is_filelike(test_set):
             test_data = self._read_data(test_set)
         else:  # test_set is a list of tuples
             test_data = test_set
@@ -245,12 +243,12 @@ class NLTKClassifier(BaseClassifier):
         return nltk.classify.accuracy(self.classifier, test_features)
 
     def update(self, new_data, *args, **kwargs):
-        '''Update the classifier with new training data and re-trains the
+        """Update the classifier with new training data and re-trains the
         classifier.
 
         :param new_data: New data as a list of tuples of the form
             ``(text, label)``.
-        '''
+        """
         self.train_set += new_data
         self.train_features = [(self.extract_features(d), c)
                                 for d, c in self.train_set]
@@ -264,8 +262,7 @@ class NLTKClassifier(BaseClassifier):
 
 
 class NaiveBayesClassifier(NLTKClassifier):
-
-    '''A classifier based on the Naive Bayes algorithm, as implemented in
+    """A classifier based on the Naive Bayes algorithm, as implemented in
     NLTK.
 
     :param train_set: The training set, either a list of tuples of the form
@@ -278,12 +275,12 @@ class NaiveBayesClassifier(NLTKClassifier):
         file format.
 
     .. versionadded:: 0.6.0
-    '''
+    """
 
     nltk_class = nltk.classify.NaiveBayesClassifier
 
     def prob_classify(self, text):
-        '''Return the label probability distribution for classifying a string
+        """Return the label probability distribution for classifying a string
         of text.
 
         Example:
@@ -297,30 +294,29 @@ class NaiveBayesClassifier(NLTKClassifier):
             0.7
 
         :rtype: nltk.probability.DictionaryProbDist
-        '''
+        """
         text_features = self.extract_features(text)
         return self.classifier.prob_classify(text_features)
 
     def informative_features(self, *args, **kwargs):
-        '''Return the most informative features as a list of tuples of the
+        """Return the most informative features as a list of tuples of the
         form ``(feature_name, feature_value)``.
 
         :rtype: list
-        '''
+        """
         return self.classifier.most_informative_features(*args, **kwargs)
 
     def show_informative_features(self, *args, **kwargs):
-        '''Displays a listing of the most informative features for this
+        """Displays a listing of the most informative features for this
         classifier.
 
         :rtype: None
-        '''
+        """
         return self.classifier.show_most_informative_features(*args, **kwargs)
 
 
 class DecisionTreeClassifier(NLTKClassifier):
-
-    '''A classifier based on the decision tree algorithm, as implemented in
+    """A classifier based on the decision tree algorithm, as implemented in
     NLTK.
 
     :param train_set: The training set, either a list of tuples of the form
@@ -333,31 +329,30 @@ class DecisionTreeClassifier(NLTKClassifier):
         file format.
 
     .. versionadded:: 0.6.2
-    '''
+    """
 
     nltk_class = nltk.classify.decisiontree.DecisionTreeClassifier
 
     def pprint(self, *args, **kwargs):
-        '''Return a string containing a pretty-printed version of this decision
+        """Return a string containing a pretty-printed version of this decision
         tree. Each line in the string corresponds to a single decision tree node
         or leaf, and indentation is used to display the structure of the tree.
 
         :rtype: str
-        '''
+        """
         return self.classifier.pp(*args, **kwargs)
 
     def pseudocode(self, *args, **kwargs):
-        '''Return a string representation of this decision tree that expresses
+        """Return a string representation of this decision tree that expresses
         the decisions it makes as a nested set of pseudocode if statements.
 
         :rtype: str
-        '''
+        """
         return self.classifier.pseudocode(*args, **kwargs)
 
 
 class PositiveNaiveBayesClassifier(NLTKClassifier):
-
-    '''A variant of the Naive Bayes Classifier that performs binary
+    """A variant of the Naive Bayes Classifier that performs binary
     classification with partially-labeled training sets, i.e. when only
     one class is labeled and the other is not. Assuming a prior distribution
     on the two labels, uses the unlabeled set to estimate the frequencies of
@@ -394,7 +389,7 @@ class PositiveNaiveBayesClassifier(NLTKClassifier):
         label ``True``.
 
     .. versionadded:: 0.7.0
-    '''
+    """
 
     nltk_class = nltk.classify.PositiveNaiveBayesClassifier
 
@@ -418,14 +413,14 @@ class PositiveNaiveBayesClassifier(NLTKClassifier):
 
     # Override
     def train(self, *args, **kwargs):
-        '''Train the classifier with a labeled and unlabeled feature sets and return
+        """Train the classifier with a labeled and unlabeled feature sets and return
         the classifier. Takes the same arguments as the wrapped NLTK class.
         This method is implicitly called when calling ``classify`` or
         ``accuracy`` methods and is included only to allow passing in arguments
         to the ``train`` method of the wrapped NLTK class.
 
         :rtype: A classifier
-        '''
+        """
         self.classifier = self.nltk_class.train(self.positive_features,
                                                 self.unlabeled_features,
                                                 self.positive_prob_prior)
@@ -434,12 +429,12 @@ class PositiveNaiveBayesClassifier(NLTKClassifier):
     def update(self, new_positive_data=None,
                new_unlabeled_data=None, positive_prob_prior=0.5,
                *args, **kwargs):
-        '''Update the classifier with new data and re-trains the
+        """Update the classifier with new data and re-trains the
         classifier.
 
         :param new_positive_data: List of new, labeled strings.
         :param new_unlabeled_data: List of new, unlabeled strings.
-        '''
+        """
         self.positive_prob_prior = positive_prob_prior
         if new_positive_data:
             self.positive_set += new_positive_data
@@ -461,7 +456,7 @@ class MaxEntClassifier(NLTKClassifier):
     nltk_class = nltk.classify.maxent.MaxentClassifier
 
     def prob_classify(self, text):
-        '''Return the label probability distribution for classifying a string
+        """Return the label probability distribution for classifying a string
         of text.
 
         Example:
@@ -475,6 +470,6 @@ class MaxEntClassifier(NLTKClassifier):
             0.7
 
         :rtype: nltk.probability.DictionaryProbDist
-        '''
+        """
         feats = self.extract_features(text)
         return self.classifier.prob_classify(feats)